利便性の都合から、blocksは純粋なブロックリストにし、block.jsonの出力はbuild.tsに統合
